chore: move monaco stuff
