+++ src/features/Editor/completion/CompletionAddProvider.ts
