+++ src/features/Editor/snippets/compSnippets.ts
